PHPSTAN Lebel 5

+++ Classes/Controller/DummyImageController.php
@@ -198,7 +198,7 @@ class DummyImageController extends ActionController
          */
         $transformedShape = array_map(
             static function (Point $point) use ($factor, $xOffset, $yOffset) {
-                return new Point($point->getX() * $factor + $xOffset, $point->getY() * $factor + $yOffset);
+                return new Point((int)($point->getX() * $factor + $xOffset), (int)($point->getY() * $factor + $yOffset));
             },
             $baseShape
         );
@@ -272,15 +272,16 @@ class DummyImageController extends ActionController
         $correctedFontSizeByHeight = $targetFontHeight * $initialFontSize / $initialFontBox->getHeight();
 
         // render actual text
-        $actualFont = $this->imagineService->font($fontFile, min([$correctedFontSizeByWidth, $correctedFontSizeByHeight]), $textColor);
+        $actualFont = $this->imagineService->font($fontFile, (int) min([$correctedFontSizeByWidth, $correctedFontSizeByHeight]), $textColor);
         $actualFontBox = $actualFont->box($text);
         $imageCenterPosition = new Point($width / 2, $height / 2);
         $textCenterPosition = new Point\Center($actualFontBox);
         if ($center) {
-            $centeredTextPosition = new Point($imageCenterPosition->getX() - $textCenterPosition->getX(), $height * .5 - $actualFontBox->getHeight() * .5);
+            $centeredTextPosition = new Point($imageCenterPosition->getX() - $textCenterPosition->getX(), (int)($height * .5 - $actualFontBox->getHeight() * .5));
         } else {
-            $centeredTextPosition = new Point($imageCenterPosition->getX() - $textCenterPosition->getX(), $height * .78 - $actualFontBox->getHeight() * .5);
+            $centeredTextPosition = new Point($imageCenterPosition->getX() - $textCenterPosition->getX(), (int)($height * .78 - $actualFontBox->getHeight() * .5));
         }
+        /** @phpstan-ignore-next-line */
         $image->draw()->text($text, $actualFont, $centeredTextPosition);
     }
